+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added:
+
+- Quick Start example and guide, [PR-54](https://github.com/reductstore/reduct-js/pull/54)
+
 ### [1.2.0] - 2022-12-23
 
 ### Added: