+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2023-02-09
+
 ### Added:
 
 - Quick Start example and guide, [PR-54](https://github.com/reductstore/reduct-js/pull/54)
@@ -131,7 +133,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/reduct-storage/reduct-js/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/reduct-storage/reduct-js/compare/v1.3.0...HEAD
+
+[1.3.0]: https://github.com/reduct-storage/reduct-js/compare/v1.2.0...v1.3.0
 
 [1.2.0]: https://github.com/reduct-storage/reduct-js/compare/v1.1.1...v1.2.0